+++ changelog.md
@@ -5,12 +5,16 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 - New method `delete_event` in customer-api v3.6.
+- New methods in configuration-api v3.6 for greetings: `create_greeting`, `delete_greeting`, `get_greeting`, `update_greeting`, `list_greetings`.
 
 ### Changed
 - Improved websocket response collection + extended logging in the websocket client.
 
 ### Bugfixes
 - Fixed version in websocket url for customer-api v3.4 and v3.6.
+
+### Removed
+- Removed support for billing-api.
 
 ## [0.4.1] - 2025-04-23